Split last update check out from config
Avoids updates to the config file for internal state tracking
