+++ internal/pkg/status/status.go
@@ -1,9 +1,10 @@
-package config
+package status
 
 import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"time"
 
 	viperlib "github.com/spf13/viper"
 )
@@ -14,21 +15,17 @@ var viper *viperlib.Viper = viperlib.New()
 // EnsureInitialised reads the config. Will quit if config is invalid
 func EnsureInitialised() {
 	if !initialised {
-		viper.SetConfigName("devcontainer-cli")
+		viper.SetConfigName("devcontainer-cli-status")
 		viper.SetConfigType("json")
 
 		viper.AddConfigPath(getConfigPath())
-
-		viper.SetDefault("templatePaths", []string{})
-		viper.SetDefault("settingPaths", []string{})
-		viper.SetDefault("experimental", false)
 
 		// TODO - allow env var for config
 		if err := viper.ReadInConfig(); err != nil {
 			if _, ok := err.(viperlib.ConfigFileNotFoundError); ok {
 				// Config file not found; ignore error if desired
 			} else {
-				fmt.Printf("Error loading config file: %s\n", err)
+				fmt.Printf("Error loading status file: %s\n", err)
 				os.Exit(1)
 			}
 		}
@@ -36,7 +33,7 @@ func EnsureInitialised() {
 	}
 }
 func getConfigPath() string {
-	path := os.Getenv("DEVCONTAINERX_CONFIG_PATH")
+	path := os.Getenv("DEVCONTAINERX_STATUS_PATH")
 	if path != "" {
 		return path
 	}
@@ -49,30 +46,27 @@ func getConfigPath() string {
 	return os.ExpandEnv(path)
 }
 
-func GetTemplateFolders() []string {
+func GetLastUpdateCheck() time.Time {
 	EnsureInitialised()
-	return viper.GetStringSlice("templatePaths")
+	return viper.GetTime("lastUpdateCheck")
 }
-func GetSnippetFolders() []string {
+func SetLastUpdateCheck(t time.Time) {
 	EnsureInitialised()
-	return viper.GetStringSlice("snippetPaths")
-}
-func GetExperimentalFeaturesEnabled() bool {
-	EnsureInitialised()
-	return viper.GetBool("experimental")
+	viper.Set("lastUpdateCheck", t)
 }
 func GetAll() map[string]interface{} {
 	EnsureInitialised()
 	return viper.AllSettings()
 }
 
-func SaveConfig() error {
+func SaveStatus() error {
 	EnsureInitialised()
 	configPath := getConfigPath()
 	configPath = os.ExpandEnv(configPath)
 	if err := os.MkdirAll(configPath, 0755); err != nil {
 		return err
 	}
-	configFilePath := filepath.Join(configPath, "devcontainer-cli.json")
+	configFilePath := filepath.Join(configPath, "devcontainer-cli-status.json")
+	fmt.Printf("HERE: %q\n", configFilePath)
 	return viper.WriteConfigAs(configFilePath)
 }
